Address two R CMD check NOTES.
A number of files that should not be included in the built-packages were added to .buildignore. The best way to avoid this kind of problem is to use dedicated functions in the usethis package, e.g. `usethis::use_travis()` will not only add .travis.yml but also will automatically add .travis.yml to .Rbuildignore.

Notes addressed:

> checking for hidden files and directories ... NOTE
  Found the following hidden files and directories:
    .travis.yml
  These were most likely included in error. See section 'Package
  structure' in the 'Writing R Extensions' manual.

> checking top-level files ... NOTE
  Non-standard files/directories found at top level:
    'CONTRIBUTING.md' 'data-raw'